Common user home page - list view, map view

+++ src/WeatherMonitorCore/WeatherMonitorCore.MqttDataSubscriberService.Interfaces/Models/SensorData.cs
@@ -2,11 +2,11 @@
 
 public readonly record struct SensorData
 {
-    public float Humidity { get; init; }
-    public float Temperature { get; init; }
-    public float AirPressure { get; init; }
-    public float Altitude { get; init; }
-    public float? PM1_0 { get; init; }
-    public float? PM2_5 { get; init; }
-    public float? PM10 { get; init; }
+    public double Humidity { get; init; }
+    public double Temperature { get; init; }
+    public double AirPressure { get; init; }
+    public double Altitude { get; init; }
+    public double? PM1_0 { get; init; }
+    public double? PM2_5 { get; init; }
+    public double? PM10 { get; init; }
 }
